+++ src/astronomicalObject/interfaces/IAstronomicalObject.ts
@@ -1,5 +1,24 @@
 import TimeOfInterest from '../../time/TimeOfInterest';
+import IRectangularCoordinates from '../../coordinates/interfaces/IRectangularCoordinates';
+import IEclipticSphericalCoordinates from '../../coordinates/interfaces/IEclipticSphericalCoordinates';
+import IEquatorialSphericalCoordinates from '../../coordinates/interfaces/IEquatorialSphericalCoordinates';
 
 export default interface IAstronomicalObject {
     toi: TimeOfInterest,
+
+    getGeocentricEclipticRectangularJ2000Coordinates(): Promise<IRectangularCoordinates>;
+
+    getGeocentricEclipticRectangularDateCoordinates(): Promise<IRectangularCoordinates>;
+
+    getGeocentricEclipticSphericalJ2000Coordinates(): Promise<IEclipticSphericalCoordinates>;
+
+    getGeocentricEclipticSphericalDateCoordinates(): Promise<IEclipticSphericalCoordinates>;
+
+    getGeocentricEquatorialSphericalJ2000Coordinates(): Promise<IEquatorialSphericalCoordinates>;
+
+    getGeocentricEquatorialSphericalDateCoordinates(): Promise<IEquatorialSphericalCoordinates>;
+
+    getApparentGeocentricEclipticSphericalCoordinates(): Promise<IEclipticSphericalCoordinates>;
+
+    getApparentGeocentricEquatorialSphericalCoordinates(): Promise<IEquatorialSphericalCoordinates>;
 }