+++ src/earth/Earth.test.js
@@ -2,24 +2,24 @@ import {createTimeOfInterest} from '../time';
 import {round} from '../utils/math';
 import Earth from './Earth';
 
-it('test getHeliocentricRectangularJ2000Coordinates', async () => {
+it('test getHeliocentricEclipticRectangularJ2000Coordinates', async () => {
     const toi = createTimeOfInterest.fromTime(2017, 12, 10, 0, 0, 0);
     const earth = new Earth(toi);
 
-    const coords = await earth.getHeliocentricRectangularJ2000Coordinates();
+    const coords = await earth.getHeliocentricEclipticRectangularJ2000Coordinates();
 
-    expect(round(coords.x, 8)).toBe(0.2070104);
+    expect(round(coords.x, 8)).toBe(0.20701041);
     expect(round(coords.y, 8)).toBe(0.96282379);
     expect(round(coords.z, 8)).toBe(-0.00004247);
 });
 
-it('test getHeliocentricRectangularDateCoordinates', async () => {
+it('test getHeliocentricEclipticRectangularDateCoordinates', async () => {
     const toi = createTimeOfInterest.fromTime(2017, 12, 10, 0, 0, 0);
     const earth = new Earth(toi);
 
-    const coords = await earth.getHeliocentricRectangularDateCoordinates();
+    const coords = await earth.getHeliocentricEclipticRectangularDateCoordinates();
 
-    expect(round(coords.x, 8)).toBe(0.20279685);
+    expect(round(coords.x, 8)).toBe(0.20279686);
     expect(round(coords.y, 8)).toBe(0.96372008);
     expect(round(coords.z, 8)).toBe(-0.00000252);
 });