+++ src/planets/index.ts
@@ -7,5 +7,13 @@ import {
     createUranus,
     createVenus
 } from './createPlanet';
+import Mercury from './Mercury';
+import Venus from './Venus';
+import Mars from './Mars';
+import Jupiter from './Jupiter';
+import Saturn from './Saturn';
+import Uranus from './Uranus';
+import Neptune from './Neptune';
 
 export {createMercury, createVenus, createMars, createJupiter, createSaturn, createUranus, createNeptune};
+export {Mercury, Venus, Mars, Jupiter, Saturn, Uranus, Neptune};