+++ src/planets/vsop87/vsop87UranusRectangularJ2000.js
@@ -2473,7 +2473,6 @@ const VSOP87_X4 = [
 
 const VSOP87_X5 = [
     [0.00000000000],
-
 ];
 
 const VSOP87_Y0 = [
@@ -4946,7 +4945,6 @@ const VSOP87_Y4 = [
 
 const VSOP87_Y5 = [
     [0.00000000000],
-
 ];
 
 const VSOP87_Z0 = [